refactor(pm): reorganize demand driver into select + driver + builder
Pure code movement (no logic change) to give each piece one job:

- `demand/select.rs` (was plan.rs) — per-edge resolution decision: pure,
  reads the store, returns `EdgeStep`. `plan_edge` -> `select_edge`.
- `demand/driver.rs` — the BFS loop + the fetch pipeline (schedule + spawn
  + pump + apply) + the `handle_processed` reporting glue, one cohesive flow.
- graph-build helpers (`process_dependency_with_resolved`,
  `try_reuse_dependency`, `chain_err`, `handle_resolved_registry_manifest`)
  move down to `resolver::builder`, beside the other build primitives.

Flattens the `demand/driver/` dir; state/queue/select/driver are now peers.
No behaviour change: 174 + 10 tests pass, clippy clean (default + all-features).
